Restructured, added CMake install target
