+++ src/Tensor.cpp
@@ -1,11 +1,10 @@
-#include <utility>
-
 //
 // Created by sergio on 13/05/19.
 //
 
-#include "../include/Tensor.h"
+#include "Tensor.h"
 
+#include <utility>
 
 Tensor::Tensor(const Model& model, const std::string& operation) {
